install filefield_paths module to allow images to be stored in directories for the node
The default entity_token implementation does not allow node tokens in image and file field paths.

There is a workaround - install the filefield_path module.  This is in beta...

    drush dl filefield_paths
    drush en -y filefield_paths

Then we can set the file path to:

    [node:content-type:machine-name]/[node:nid]/images

And the file name to:

    [file:ffp-name-only-original].[file:ffp-extension-original]

Which does more or less what I want. I also enabled the "transliterate"
option on the file name.

Somewhat annoyingly, this is only *almost* what I want. The main image goes where I expect:

    sites/default/files/announcement/882/images/summer-fair-report_2012-08-25-1.jpg

Resized images go elsewhere:

    sites/default/files/styles/medium/public/announcement/882/images/summer-fair-report_2012-08-25-1.jpg

Ignoring this for the moment, as I can't see any way around it.